refactor: properly implement state pattern in KeyControls:
the KeyControls only has an active controller, the controller returns the new state of the controller after its excution. This makes the control flow much more natural. Move some logic from KeyControls to the ManfredController, this should be removed from there in a later step again.
still to be done:
- repair GelaberController
- introduce Command pattern for callbacks from Interactable to ManfredController and remove weird methods from ManfredController